# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nuclom API",
    "version": "1.0.0",
    "description": "Nuclom is a video collaboration platform that helps teams organize, share, and collaborate on video content. This API provides programmatic access to videos, organizations, series, comments, and more.",
    "contact": {
      "name": "Nuclom Support",
      "url": "https://nuclom.com/support"
    },
    "license": {
      "name": "Elastic License 2.0",
      "url": "https://www.elastic.co/licensing/elastic-license"
    }
  },
  "servers": [
    {
      "url": "https://nuclom.com/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/ai/analyze": {
      "post": {
        "tags": [
          "AI"
        ],
        "summary": "Analyze video",
        "operationId": "createAiAnalyze",
        "description": "Trigger AI analysis of a video (transcript, summary, action items, tags)",
        "responses": {
          "200": {
            "description": "Analysis result",
            "content": {
              "application/json": {
                "schema": {
                  "$defs": {
                    "UUID": {
                      "type": "string",
                      "description": "a Universally Unique Identifier",
                      "format": "uuid",
                      "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                    }
                  },
                  "type": "object",
                  "required": [
                    "videoId",
                    "processingTime"
                  ],
                  "properties": {
                    "videoId": {
                      "$ref": "#/$defs/UUID"
                    },
                    "transcript": {
                      "type": "string"
                    },
                    "summary": {
                      "type": "string"
                    },
                    "actionItems": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "processingTime": {
                      "type": "number"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/channels": {
      "get": {
        "tags": [
          "Channels"
        ],
        "summary": "List channels",
        "operationId": "listChannel",
        "description": "Get channels for an organization",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Channels"
        ],
        "summary": "Create channel",
        "operationId": "createChannel",
        "description": "Create a new channel",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/integrations/{integrationId}/settings": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "summary": "Get setting",
        "operationId": "getIntegrationSetting",
        "parameters": [
          {
            "name": "integrationId",
            "in": "path",
            "description": "IntegrationId ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Integrations"
        ],
        "summary": "Patch setting",
        "operationId": "patchIntegrationSetting",
        "parameters": [
          {
            "name": "integrationId",
            "in": "path",
            "description": "IntegrationId ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Authentication token from login or API key"
      }
    }
  }
}